removing CommandList and migrate structures to Commands
