+++ yakshieva.rozay/T3/Commands.cpp
@@ -7,7 +7,6 @@
 #include <numeric>
 #include <limits>
 #include "AreaCalculator.hpp"
-#include "CommandList.hpp"
 #include "StreamGuard.hpp"
 
 void yakshieva::doAreaCommand(const std::vector< Polygon >& polygons, std::istream& in, std::ostream& out)
@@ -119,14 +118,44 @@ void yakshieva::doCountCommand(const std::vector< Polygon >& polygons, std::istr
 
 void yakshieva::doLessAreaCommand(const std::vector< Polygon >& polygons, std::istream& in, std::ostream& out)
 {
- Polygon refPolygon;
- in >> refPolygon;
- if (!(in && in.peek() == '\n')) {
+  Polygon refPolygon;
+  in >> refPolygon;
+  if (!(in && in.peek() == '\n'))
+  {
     throw std::invalid_argument("<INVALID COMMAND>");
   }
   using namespace std::placeholders;
-auto lessAreaCount = std::count_if(polygons.begin(), polygons.end(), LessArea{ accamulateAllParts(refPolygon) });
+  auto lessAreaCount = std::count_if(polygons.begin(), polygons.end(), LessArea{ accamulateAllParts(refPolygon) });
   out << lessAreaCount << "\n";
+}
+
+namespace yakshieva
+{
+  struct PolygonComparator
+  {
+    const Polygon& target;
+    bool operator()(const Polygon& p1, const Polygon& p2) const;
+  };
+  size_t removeConsecutiveDuplicates(std::vector< Polygon >& polygons, const Polygon& target);
+}
+
+bool yakshieva::PolygonComparator::operator()(const Polygon& p1, const Polygon& p2) const
+{
+  return p1 == p2 && p2 == target;
+}
+
+size_t yakshieva::removeConsecutiveDuplicates(std::vector< Polygon >& polygons, const Polygon& target)
+{
+  PolygonComparator comparator{ target };
+  std::size_t removedCount = 0;
+  auto it = std::adjacent_find(polygons.begin(), polygons.end(), std::bind(comparator, std::placeholders::_1, std::placeholders::_2));
+  while (it != polygons.end())
+  {
+    it = polygons.erase(it);
+    ++removedCount;
+    it = std::adjacent_find(it, polygons.end(), std::bind(comparator, std::placeholders::_1, std::placeholders::_2));
+  }
+  return removedCount;
 }
 
 void yakshieva::doRmechoCommand(std::vector< Polygon >& polygons, std::istream& in, std::ostream& out)
@@ -143,6 +172,19 @@ void yakshieva::doRmechoCommand(std::vector< Polygon >& polygons, std::istream& 
   out << removedCount << "\n";
 }
 
+void yakshieva::duplicateEntries(std::vector<Polygon>& p, const Polygon& targetP)
+{
+  std::vector<Polygon> temp;
+  for (const auto& poly : p)
+  {
+    temp.push_back(poly);
+    if (poly == targetP)
+    {
+      temp.push_back(targetP);
+    }
+  }
+  p = std::move(temp);
+}
 
 void yakshieva::doEchoCommand(std::vector< Polygon >& polygons, std::istream& in, std::ostream& out)
 {
